Adding project as is
This adds the SkyCRM project which I had worked on between 2020 and 2023. To date, it's still handling EverBright webhook matters, but much of it has had pieces of it replaced.

When this application was most active, it would handle passing information between Copper, Hubspot, Netsuite, OneDrive/SharePoint, PandaDocs, EverBright (Formerly Sighten), Google, Slack, and Zenefits.

In general, it would receive HTTP messages such as POST messages. These messages would be stored in Postgres using Oban. Oban would then at some point quickly after perform the message processing instructions defined in the Oban.Worker modules.

Phoenix and LiveView were used to provide a dashboard for reporting what the system was doing and what tasks it had completed and failed at. This would be done by querying on the Oban.Job schema stored in Postgres.

The application would be built into a docker image and that would run on a Digital Ocean droplet. Let's Encrypt was used to provide SSL.

This has definitely been my largest Elixir project and I'm displaying it to show off to employers.